+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.6.3] - 2026-01-13
+## [0.6.3] - xxxx-xx-xx
 
 ### Changes
 
@@ -14,7 +14,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add CLI commands:
   - `image`
     - `upload`
-- Python: Accept `collections::abc::Buffer` instead of `bytes` wherever possible
 
 ## [0.6.2] - 2026-01-13